handle product exist and not exist in consumer

+++ consumer.py
@@ -1,4 +1,5 @@
 import pika, json, os
+from sqlalchemy.exc import IntegrityError
 
 from main import Product, db
 
@@ -10,26 +11,33 @@ channel = connection.channel()
 
 channel.queue_declare(queue='main')
 
+
 def callback(ch, method, properties, body):
     print('Received in main')
     data = json.loads(body)
     print(data)
 
     if properties.content_type == 'product_created':
-        product = Product(id=data['id'], title=data['title'], image=data['image'])
-        db.session.add(product)
-        db.session.commit()
+        try:
+            product = Product(id=data['id'], title=data['title'], image=data['image'])
+            db.session.add(product)
+            db.session.commit()
+        except IntegrityError:
+            db.session.rollback()
 
     elif properties.content_type == 'product_updated':
         product = Product.query.get(data['id'])
-        product.title = data['title']
-        product.image = data['image']
-        db.session.commit()
+        if product is not None:
+            product.title = data['title']
+            product.image = data['image']
+            db.session.commit()
 
     elif properties.content_type == 'product_deleted':
         product = Product.query.get(data)
-        db.session.delete(product)
-        db.session.commit()
+        if product is not None:
+            db.session.delete(product)
+            db.session.commit()
+
 
 channel.basic_consume(queue='main', on_message_callback=callback, auto_ack=True)
